requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-jalali==7.3.0
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.15.2
djangorestframework-dataclasses==1.3.1
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-environ==0.4.54
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
polars-lts-cpu==1.29.0
redis>=4.6.0
django-celery-beat>=2.5.0
